security

+++ TaskListSuper2021/Pages/Delete.cshtml.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -35,6 +36,12 @@ namespace TaskListSuper2021.Pages
                 .Include(t => t.Author)
                 .Include(t => t.Receiver).FirstOrDefaultAsync(m => m.TaskId == id);
 
+            var userId = User.Claims.Where(c => c.Type == ClaimTypes.NameIdentifier).FirstOrDefault().Value;
+            if(TaskItem.AuthorId != userId)
+            {
+                return Unauthorized();
+            }
+
             if (TaskItem == null)
             {
                 return NotFound();
@@ -50,7 +57,11 @@ namespace TaskListSuper2021.Pages
             }
 
             TaskItem = await _context.Tasks.FindAsync(id);
-
+            var userId = User.Claims.Where(c => c.Type == ClaimTypes.NameIdentifier).FirstOrDefault().Value;
+            if (TaskItem.AuthorId != userId)
+            {
+                return Unauthorized();
+            }
             if (TaskItem != null)
             {
                 _context.Tasks.Remove(TaskItem);
